Address review: simplify to the ExplicitOuter fix only
- Drop the Inliner this-proxy change: the outer-accessor reconstruction in
  ExplicitOuter handles every case on its own, so the inliner no longer needs
  to special-case naming the enclosing `this`.
- Drop the `assumeInline` parameter of `referencesOuter`: a module's outer
  accessor is only needed downstream when inlining emits an outer *select*,
  which only happens for value-level outer references; those are always caught
  without `inInline`. Type-only references (the only thing `inInline` adds) are
  erased and never produce a downstream outer select.
- Shorten the `referencesOuterViaInlineMembers` doc comment.
- Replace the macro reproducer with the simpler non-macro `tests/pos/i26352`
  (no macro/`transparent` needed to trigger the crash).
