+++ tests/pos/i26352/Outer_1.scala
@@ -1,0 +1,7 @@
+// https://github.com/scala/scala3/issues/26352
+class Box[I]
+
+abstract class Outer:
+  self =>
+  object inner:
+    inline def make: Box[?] = new Box[self.type]
